+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './common';
 export * from './decorators';
 export * from './metadata';
+export * from './FactoryLogger';
 export * from './FixtureFactory';