+++ test/factory.test.ts
@@ -1,8 +1,32 @@
-import { FixtureFactory } from '../src/FixtureFactory';
+import { FixtureFactory, Assigner } from '../src/FixtureFactory';
 import { Fixture } from '../src/decorators/Fixture';
+import {
+  IsIn,
+  Equals,
+  IsPositive,
+  Min,
+  Max,
+  MinDate,
+  MaxDate,
+  Contains,
+  IsAlpha,
+  IsAlphanumeric,
+  IsEmail,
+  IsFQDN,
+  IsHexColor,
+  IsLowercase,
+  IsUppercase,
+  Length,
+  MinLength,
+  MaxLength,
+  ArrayContains,
+  ArrayMinSize,
+  ArrayMaxSize,
+  IsNegative,
+} from 'class-validator';
 
 describe(`FixtureFactory`, () => {
-  const factory = new FixtureFactory({ logging: true });
+  const factory = new FixtureFactory({ logging: false });
 
   it(`makes metadata of registered entities`, () => {
     class DummyAuthor {}
@@ -225,6 +249,22 @@ describe(`FixtureFactory`, () => {
       expect(person.book.author).toBeUndefined();
     });
 
+    it(`class [many to many]`, () => {
+      class Book {
+        @Fixture({ type: () => [BookTag] })
+        tags!: BookTag[];
+      }
+      class BookTag {
+        @Fixture({ type: () => [Book] })
+        books!: Book[];
+      }
+      factory.register([BookTag, Book]);
+
+      const book = factory.make(Book).one();
+      expect(book.tags[0]).toBeInstanceOf(BookTag);
+      expect(book.tags[0].books).toBeUndefined();
+    });
+
     it(`multi-level nesting`, () => {
       class Author {
         @Fixture()
@@ -254,6 +294,280 @@ describe(`FixtureFactory`, () => {
       expect(author.books[0]).toBeDefined();
       expect(author.books[0].tags[0]).toBeDefined();
       expect(author.books[0].tags[0].category).toBeDefined();
+    });
+
+    it(`custom assigner`, () => {
+      const factory = new FixtureFactory({ logging: true });
+      const assigner: Assigner = (prop, obj, _value) => {
+        obj[prop.name] = 'foo';
+      };
+      const mock = jest.fn((...args: Parameters<Assigner>) =>
+        assigner(...args)
+      );
+      class Author {
+        @Fixture()
+        name!: string;
+      }
+      factory.register([Author]);
+      factory.setAssigner(mock);
+
+      const author = factory.make(Author).one();
+
+      expect(mock).toHaveBeenCalled();
+      expect(author.name).toBe('foo');
+    });
+  });
+
+  describe(`with class-validator`, () => {
+    it(`@IsIn()`, () => {
+      class Dummy {
+        @IsIn(['a', 'b', 'c'])
+        val!: string;
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      expect(['a', 'b', 'c'].includes(dummy.val)).toBe(true);
+    });
+
+    it(`@Equals()`, () => {
+      class Dummy {
+        @Equals('foo')
+        val!: string;
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      expect(dummy.val).toBe('foo');
+    });
+
+    it(`@IsPositive()`, () => {
+      class Dummy {
+        @IsPositive()
+        val!: number;
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      expect(dummy.val > 0).toBe(true);
+    });
+
+    it(`@IsNegative()`, () => {
+      class Dummy {
+        @IsNegative()
+        val!: number;
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      expect(dummy.val < 0).toBe(true);
+    });
+
+    it(`@Min() `, () => {
+      class Dummy {
+        @Min(500)
+        val!: number;
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      expect(dummy.val >= 500).toBe(true);
+    });
+
+    it(`@Max() `, () => {
+      class Dummy {
+        @Max(500)
+        val!: number;
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      expect(dummy.val <= 500).toBe(true);
+    });
+
+    it(`@MinDate() `, () => {
+      const date = new Date();
+      class Dummy {
+        @MinDate(date)
+        val!: Date;
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      expect(+dummy.val > +date).toBe(true);
+    });
+
+    it(`@MaxDate() `, () => {
+      const date = new Date();
+      class Dummy {
+        @MaxDate(date)
+        val!: Date;
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      expect(+dummy.val < +date).toBe(true);
+    });
+
+    it(`@Contains() `, () => {
+      class Dummy {
+        @Contains('foo')
+        val!: string;
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      expect(dummy.val.includes('foo')).toBe(true);
+    });
+
+    it(`@IsAlpha() `, () => {
+      class Dummy {
+        @IsAlpha()
+        val!: string;
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      expect(dummy.val).not.toMatch(/\d/);
+    });
+
+    it(`@IsAlphanumeric() `, () => {
+      class Dummy {
+        @IsAlphanumeric()
+        val!: string;
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      expect(typeof dummy.val).toBe('string');
+    });
+
+    it(`@IsEmail() `, () => {
+      class Dummy {
+        @IsEmail()
+        val!: string;
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      expect(dummy.val).toMatch(/@.+\..+/);
+    });
+
+    it(`@IsFQDN() `, () => {
+      class Dummy {
+        @IsFQDN()
+        val!: string;
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      expect(dummy.val).toMatch(/.+\..+/);
+    });
+
+    it(`@IsHexColor() `, () => {
+      class Dummy {
+        @IsHexColor()
+        val!: string;
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      expect(dummy.val).toMatch(/^#/);
+    });
+
+    it(`@IsLowercase() `, () => {
+      class Dummy {
+        @IsLowercase()
+        val!: string;
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      expect(dummy.val.toLowerCase()).toBe(dummy.val);
+    });
+
+    it(`@IsUppercase() `, () => {
+      class Dummy {
+        @IsUppercase()
+        val!: string;
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      expect(dummy.val.toUpperCase()).toBe(dummy.val);
+    });
+
+    it(`@Length() `, () => {
+      class Dummy {
+        @Length(20, 30)
+        val!: string;
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      console.log('[Length]: ', dummy.val, dummy.val.length);
+      expect(dummy.val.length >= 20 && dummy.val.length <= 30).toBe(true);
+    });
+
+    it(`@MinLength() `, () => {
+      class Dummy {
+        @MinLength(20)
+        val!: string;
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      console.log('[MinLength]: ', dummy.val, dummy.val.length);
+      expect(dummy.val.length >= 20).toBe(true);
+    });
+
+    it(`@MaxLength() `, () => {
+      class Dummy {
+        @MaxLength(5)
+        val!: string;
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      console.log('[MaxLength]: ', dummy.val, dummy.val.length);
+      expect(dummy.val.length <= 5).toBe(true);
+    });
+
+    it(`@ArrayContains() `, () => {
+      class Dummy {
+        @ArrayContains([1, 2])
+        val!: any[];
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      expect(dummy.val).toEqual([1, 2]);
+    });
+
+    it(`@ArrayMinSize() `, () => {
+      class Dummy {
+        @ArrayMinSize(4)
+        @Fixture({ type: () => [String] })
+        val!: string[];
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      console.log('[ArrayMinSize] dummy :', dummy);
+      expect(dummy.val.length >= 4).toBe(true);
+    });
+
+    it(`@ArrayMaxSize() `, () => {
+      class Dummy {
+        @ArrayMaxSize(2)
+        @Fixture({ type: () => [String] })
+        val!: string[];
+      }
+      factory.register([Dummy]);
+
+      const dummy = factory.make(Dummy).one();
+      console.log('[ArrayMaxSize] dummy :', dummy);
+      expect(dummy.val.length <= 2).toBe(true);
     });
   });
 });